runner/requirements.txt
django
djangorestframework
requests
celery
docker
tenacity
